src/previous_monlan_versions/6_monlan_dqn_pytorch_2021_2022/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
cycler==0.11.0
fonttools==4.30.0
idna==3.3
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.5.1
MetaTrader5==5.0.36
networkx==2.7.1
numpy==1.22.3
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
six==1.16.0
threadpoolctl==3.1.0
tqdm==4.63.0
typing_extensions==4.1.1
urllib3==1.26.8
